pyproject.toml
[project]
name = "asr-finetuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.0.0",
    "pytorch-lightning>=2.6.1",
    "pyyaml>=6.0.0",
    "torchcodec>=0.10.0",
    "torchmetrics[text]>=1.8.2",
    "trackio>=0.16.1",
    "transformers>=4.57.6",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["asr_finetuning*"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.ruff]
line-length = 88
